gdb: disable python when python is not available.
The patch checkes whether the python executable is available
at runtime. If it is not available, it disables python at run_time.

This backports a local patch from ChromeOS. The bug entry for
the ChromeOS is at crbug.com/248605

This CL also check some pre defines to include unistd.h and call
the function fork().

BUG: crbug.com/248605