bfd: add experimental support for SHT_RELR sections.
This change adds experimental support for SHT_RELR sections, proposed
here: https://groups.google.com/forum/#!topic/generic-abi/bX460iggiKg

This is a subset of the change that was merged into the aosp/binutils
repo (https://android-review.googlesource.com/572669). It contains the
definitions for the new section type / dynamic tags, and support in bfd
for recognizing the new section type. Gdb uses bfd to read the program
binary, and without this support, it fails to detect some architecture
related settings.

Bug: https://bugs.chromium.org/p/chromium/issues/detail?id=836845
Test: Built gdb inside a Chrome OS chroot. Verified that architecture
      auto-detection works for an x86_64 binary containing a '.relr.dyn'
      section. The message "don't know how to handle section `.relr.dyn'"
      is no longer printed.

Change-Id: Id199a7201d4750d2c6cb41a0c1a811777c322621